minor code reorganization
